refactor: decompose read.go into resolve.go and filter.go
Extract object graph resolution (resolve, resolveInStream) into resolve.go,
stream filter decoders (applyFilter, pngUpReader) into filter.go alongside
the existing ASCII85 helpers, and objfmt into object.go where the object
types it formats are defined. read.go now contains only the Reader public
API: Open, NewReader, OpenBytes, NewReaderEncrypted, Trailer.
