+++ read.go
@@ -61,13 +61,9 @@ package pdf
 
 import (
 	"bytes"
-	"compress/zlib"
-	"encoding/ascii85"
 	"fmt"
 	"io"
 	"os"
-	"sort"
-	"strconv"
 )
 
 // DebugOn is responsible for logging messages into stdout. If problems arise during reading, set it true.
@@ -161,209 +157,3 @@ func NewReaderEncrypted(f io.ReaderAt, size int64, pw func() string) (r *Reader,
 func (r *Reader) Trailer() Value {
 	return Value{r, r.trailerptr, r.trailer}
 }
-
-func objfmt(x interface{}) string {
-	switch x := x.(type) {
-	default:
-		return fmt.Sprint(x)
-	case string:
-		if isPDFDocEncoded(x) {
-			return strconv.Quote(pdfDocDecode(x))
-		}
-		if isUTF16(x) {
-			return strconv.Quote(utf16Decode(x[2:]))
-		}
-		return strconv.Quote(x)
-	case name:
-		return "/" + string(x)
-	case dict:
-		var keys []string
-		for k := range x {
-			keys = append(keys, string(k))
-		}
-		sort.Strings(keys)
-		var buf bytes.Buffer
-		buf.WriteString("<<")
-		for i, k := range keys {
-			elem := x[name(k)]
-			if i > 0 {
-				buf.WriteString(" ")
-			}
-			buf.WriteString("/")
-			buf.WriteString(k)
-			buf.WriteString(" ")
-			buf.WriteString(objfmt(elem))
-		}
-		buf.WriteString(">>")
-		return buf.String()
-
-	case array:
-		var buf bytes.Buffer
-		buf.WriteString("[")
-		for i, elem := range x {
-			if i > 0 {
-				buf.WriteString(" ")
-			}
-			buf.WriteString(objfmt(elem))
-		}
-		buf.WriteString("]")
-		return buf.String()
-
-	case stream:
-		return fmt.Sprintf("%v@%d", objfmt(x.hdr), x.offset)
-
-	case objptr:
-		return fmt.Sprintf("%d %d R", x.id, x.gen)
-
-	case objdef:
-		return fmt.Sprintf("{%d %d obj}%v", x.ptr.id, x.ptr.gen, objfmt(x.obj))
-	}
-}
-
-
-// resolveInStream locates ptr inside an ObjStm (object stream) by scanning
-// the index entries, following Extends chains as needed.
-func (r *Reader) resolveInStream(parent objptr, ptr objptr, xr xref) interface{} {
-	strm := r.resolve(parent, xr.stream)
-Search:
-	for {
-		if strm.Kind() != Stream {
-			panic("not a stream")
-		}
-		if strm.Key("Type").Name() != "ObjStm" {
-			panic("not an object stream")
-		}
-		n := int(strm.Key("N").Int64())
-		first := strm.Key("First").Int64()
-		if first == 0 {
-			panic("missing First")
-		}
-		b := newBuffer(strm.Reader(), 0)
-		b.allowEOF = true
-		for i := 0; i < n; i++ {
-			id, _ := b.readToken().(int64)
-			off, _ := b.readToken().(int64)
-			if uint32(id) == ptr.id {
-				b.seekForward(first + off)
-				return b.readObject()
-			}
-		}
-		ext := strm.Key("Extends")
-		if ext.Kind() != Stream {
-			panic("cannot find object in stream")
-		}
-		strm = ext
-		continue Search
-	}
-}
-
-func (r *Reader) resolve(parent objptr, x interface{}) Value {
-	if ptr, ok := x.(objptr); ok {
-		if ptr.id >= uint32(len(r.xref)) {
-			return Value{}
-		}
-		xr := r.xref[ptr.id]
-		if xr.ptr != ptr || !xr.inStream && xr.offset == 0 {
-			return Value{}
-		}
-		if xr.inStream {
-			x = r.resolveInStream(parent, ptr, xr)
-		} else {
-			b := newBuffer(io.NewSectionReader(r.f, xr.offset, r.end-xr.offset), xr.offset)
-			b.key = r.key
-			b.useAES = r.useAES
-			obj := b.readObject()
-			def, ok := obj.(objdef)
-			if !ok {
-				panic(fmt.Errorf("loading %v: found %T instead of objdef", ptr, obj))
-			}
-			if def.ptr != ptr {
-				panic(fmt.Errorf("loading %v: found %v", ptr, def.ptr))
-			}
-			x = def.obj
-		}
-		parent = ptr
-	}
-
-	switch x := x.(type) {
-	case nil, bool, int64, float64, name, dict, array, stream:
-		return Value{r, parent, x}
-	case string:
-		return Value{r, parent, x}
-	default:
-		panic(fmt.Errorf("unexpected value type %T in resolve", x))
-	}
-}
-
-
-func applyFilter(rd io.Reader, name string, param Value) io.Reader {
-	switch name {
-	default:
-		panic("unknown filter " + name)
-	case "FlateDecode":
-		zr, err := zlib.NewReader(rd)
-		if err != nil {
-			panic(err)
-		}
-		pred := param.Key("Predictor")
-		if pred.Kind() == Null {
-			return zr
-		}
-		columns := param.Key("Columns").Int64()
-		switch pred.Int64() {
-		default:
-			if DebugOn {
-				fmt.Println("unknown predictor", pred)
-			}
-			panic("pred")
-		case 12:
-			return &pngUpReader{r: zr, hist: make([]byte, 1+columns), tmp: make([]byte, 1+columns)}
-		}
-	case "ASCII85Decode":
-		cleanASCII85 := newAlphaReader(rd)
-		decoder := ascii85.NewDecoder(cleanASCII85)
-
-		switch param.Keys() {
-		default:
-			if DebugOn {
-				fmt.Println("param=", param)
-			}
-			panic("not expected DecodeParms for ascii85")
-		case nil:
-			return decoder
-		}
-	}
-}
-
-type pngUpReader struct {
-	r    io.Reader
-	hist []byte
-	tmp  []byte
-	pend []byte
-}
-
-func (r *pngUpReader) Read(b []byte) (int, error) {
-	n := 0
-	for len(b) > 0 {
-		if len(r.pend) > 0 {
-			m := copy(b, r.pend)
-			n += m
-			b = b[m:]
-			r.pend = r.pend[m:]
-			continue
-		}
-		_, err := io.ReadFull(r.r, r.tmp)
-		if err != nil {
-			return n, err
-		}
-		if r.tmp[0] != 2 {
-			return n, fmt.Errorf("malformed PNG-Up encoding")
-		}
-		for i, b := range r.tmp {
-			r.hist[i] += b
-		}
-		r.pend = r.hist[1:]
-	}
-	return n, nil
-}
-
